requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
azure-ai-ml==1.21.0
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.19.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.23.1
azure-storage-file-datalake==12.17.0
azure-storage-file-share==12.19.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.9.0
google-api-core==2.21.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
gradio==5.0.1
gradio_client==1.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
numpy==2.1.2
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydash==8.0.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==307
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
ruff==0.6.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sseclient==0.0.27
starlette==0.38.6
strictyaml==1.7.3
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==12.0
